Favicon, Transcripts, Browser Support, Syncronous JSON Requests
Site has an icon in the tab toolbar. New transcripts are available in docx and json format. Gitignore is now ignoring mp3's again cause there's a lot of them. Starting to log confirmed and tested browser support on the ods sheet. Added an ajax handler for getting json into a javascript object, implemented it on interviews section.