Map Tab All But Done
Adding to the map is just a matter of adding to NASCAMapPoints.json which is super self-explanatory. Leaflet popups on map can't play videos for some reason but the rest should be fine. Started working on merging the timeline but there's just not really enough working and done right now from what I can see to invest in it yet.