contextual search bar and css fix
Search bar slides in if you're in a tab that uses it. Otherwise it stays out of the window. Also, I noticed on my laptop when the site first loads dynamic_css() was not being called, at least at the right time. So I added a call to it in a few additional init functions to ensure that it fixes everything on load without having to reposition the browser window to get it to call and fix itself.