Landing Page Complete
With the exception of the custom view for letters which is yet to be designed and locked down, the landing page is complete. All functionality and styling is there by design. The description details are easy to change but for now I have it pulling the description from content dm, in addition to the tribe name and photographer.