Website CSS Sizing Fix
Turned down the minimum size (in pixels) that items and text can shrink to. It should fix the sizing problems across monitors of different pixel densities. I personally think this new minimum is way too low and will likely be bumped up to some middle ground later on, but for now I'm leaving it here so I can illustrate the issue to the team.