Landing Page - Error Messages and Default Titles
Styled the error message output when home-more.php returns with an error. Also added custom additional text for error code -4 which will commonly pop up whenever contentDM is updated and our site is not. Also styled and modified the default text when no cards are selected, welcoming you to the page and giving you a starting point on what you might want to do while you're at the page.