Home page additional functionality. See description.
If a card has an error, the script will replace it. So you should never see cards with errors on home page. Home page now indexes and accurately displays 'video' cards. Custom video right-side view added to handle this. Additional curl function in cdm.php that checks if there is a file at a given url. Needed this to double check for existence of thumbnails from youtube.